requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
pydantic>=2.6
pydantic-settings>=2.2
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
python-dotenv>=1.0
tenacity>=8.2
lxml>=5.1
trafilatura>=1.8
readability-lxml>=0.8.1
markdownify>=0.12
playwright>=1.41
google-generativeai>=0.4.0
requests>=2.31
pytest>=8.0
pytest-asyncio>=0.23
pytest-httpx>=0.30
ruff>=0.2
